updated test methods to be compartmentalized by usage
